requirements.txt
numpy==1.24.3
Pillow==10.0.0
scipy==1.11.2
matplotlib==3.7.2
rasterio==1.3.10
torch==2.2.2
torchvision==0.17.2

